Combine milestone and assignees gihub workflows
These two were using the same triggers and are in the same bucket of
work which is to be done after a PR is merged. With the changes in
this PR the previous two standalone github workflows were combined
into a single workflow with two job steps doing the previous work.