Move the "merge_job" to the pr_merge_cleanup workflow
The merge_job from label.yml was in charge of removing labels that were
 initially added to the PR and to also add comment in case there are
string changes. This type of work belongs in the pr_merge_cleanup
worflow.
This has the added benefit of removing this workflow from the pr UI
related to checks.

Also fixes two issues:
- removes a checkout of the repo that wasn't needed
- fixes a bug which resulted in the string related comment to not be
added(funny this didn't work until now, maybe it's not needed at all?)