Inline usages of StdModels declared models
This is the initial work to remove StdModels(used only by tests).

The initializer functions code used to create the models in StdModels
was moved to standalone functions to be used by interested tests. The
only difference is the addition of a call to save the model:

col.notetypes.save(m)

Note: StdModels were both used in unit and instrumented tests. As
there's only one usage in an instrumented test I duplicated the code
to avoid the use of the testlib.