Move ClosableDrawerLayout to workarounds package
Kotlin is more flexible with its package structure and we can declare
classes with different packages vs. the actual folder structure. This
avoids the IDE wrapping the packages in its Project View due to having
two root packages.

Also it's a step in having com.ichi2.anki as the root package for the
main module.