Fix color scheme for empty cards report dialog
This removes the initial implementation using an WebView in favor of a
TextView and parsing the report with HtmlCompat. The nids links are
now implemented through ClickSpans.

Also adds a test to check the html backend report structure to make
sure the report parsing code present in EmptyCardsFragment stays
relevant.