Remove StudyOptionsListener from StudyOptionsFragment
This listener was invoked only when DeckPicker was in tablet mode(and
StudyOptionsFragment was visible besides the decks list) and when
the "resetDecksList" argument was set to true.

The "resetDecksList" was set to:
- false everytime refreshInterface() was called from outside the
fragment
- true when refreshInterface() was called from inside the fragment,
when unburying a deck or rebuilding/emptying filtered decks

The "resetDecksList" set to true cases were handled by relying on
ChangeManager to update the UI as those methods are undoable.

Note: the test was changed to handle a failure for the xlarge qualifier.
For some reason the DeckPicker state is updated correctly but it
doesn't properly update the menu in tablet mode(in tests). I think this
is an issue with how we handle the menu state in DeckPicker and not a bug
due to the changes in this PR(I tested the actions that were previously
invoking the listener in both portrait and tablet and they work correctly).