Fix state restore bug in FindAndReplaceDialogFragment
To get the current selected notes the fragment is querying the browser's
ViewModel for the selected notes.

This works most of the times but fails(we lose the selection and will
target everything instead of only the selected notes) when going to
the background and coming back(and fragment gets recreated). At this point
the fragment tries to access again the selected notes but these are not
available as they are being manually restored from a file.

This bug can be seen by using the "Don't keep activities" dev option.

The fix consists in using an IdsFile so the fragment always has a
reference to the selected notes from the moment of its creation.